Preparation for upload to test flight. Updated name of app shown to Pray.
Added functionality to show whether the prayer topic has been prayed for today (in the main prayerListTableViewController). Added double-tap and single-tap gesture recognizers to allow that functionality. Added method to move rows from one section to another within that view. Added Placeholder to instruct the user that they can double-tap to mark something as having been prayed for.

Added functionality to show when you have prayed for a specific topic in the prayerRequestEditorViewController. Added mayUpdate functions so that the system is more modular.

Updated refreshing an individual prayer topic to make it work correctly. Made the functions more modular and able to handle situations where we want to update some of the dates in the date frame (the ones that are still to come) but not all of them (excluding the ones that have already passed).